+++ .gitignore
@@ -45,3 +45,7 @@ app.*.map.json
 kitx_mobile/android/app/debug
 kitx_mobile/android/app/profile
 kitx_mobile/android/app/release
+kitx_mobile/windows
+kitx_mobile/web
+kitx_mobile/macos
+kitx_mobile/linux
