+++ kitx_mobile/assets/i18n/zh_CN.json
@@ -1,6 +1,7 @@
 {
   "drawer": {
     "home": "主页",
+    "dashboard": "仪表板",
     "devices": "设备管理",
     "account": "账户",
     "setting": "设置"
